Add multi-tenancy support to ontology context and session management.
The update provides user-specific access to the ontology context by setting a UserService for each session. User graph URIs are now set based on the UserService. The commit also updates session handling in the CallbackServlet for Auth0 callbacks after the user logs in, and HomeServlet for logged in users sessions. As a result, individual users receive and can work on a session-specific duplicate of the default graph, ensuring improved isolation of user actions.

Remaining ToDos:
Proper ErrorHandling where cookies are not passed. Therefore, in certain cases an error should not be thrown.
Check if SessionValidationServlet.java can be done via JAX RS instead of Servlets, it would be easier and more consistent.
Currently, because of a bug, setUserService is overloaded. Check if possible to remove the overloaded version.
Check if CookieResponseFilter.java is needed. if not, remove.