Add authentication bypass for development environment
A new servlet `DevEnv` has been introduced to help development work by setting up a test user that bypasses normal authentication. It is checked in the `SessionValidationServlet` and executed if the app environment is set to "development". This helps developers by reducing the time spent on repeat logins during development testing.