Automatic authentication in Dev environment
The authentication flow in the development environment was updated to automatically handle authentication. The DevEnv class was refactored to remove the environment check and always perform the login operation, and getters were added for the idToken and accessToken. Also, the SessionValidationServlet was updated to adopt this new logic, implementing an auto-authentication in the development environment. This aims to simplify the process for testing in a development environment.